HQC連接符号の符号化を実装

+++ src/hqc/hqc_param.rs
@@ -4,12 +4,12 @@ pub struct HQCParam<'a> {
     pub rs_n: usize,
     pub rs_k: usize,
     pub rs_symbol_field: GaloisField2m<'a>,
-    pub rs_genpoly: Vec<u16>,
+    pub rs_genpoly: Vec<u8>,
     pub hadamard_multiplicity: u8,
 }
 
 impl<'a> HQCParam<'a> {
-    pub fn new(rs_n: usize, rs_k: usize, rs_genpoly: Vec<u16>, hadamard_multiplicity: u8) -> Self {
+    pub fn new(rs_n: usize, rs_k: usize, rs_genpoly: Vec<u8>, hadamard_multiplicity: u8) -> Self {
         Self {
             rs_n,
             rs_k,
